frontend: update to use non-object sanization
